created directory structure for deployables
